Add free-list heap, scheduler, syscalls, and freestanding libc
Expand the bare-metal core: proper kmalloc/kfree/krealloc, Normal/Device
MMU attrs, cooperative scheduler, SVC ABI, and a libc shim for higher-level
code. Clean up boot/linker stack layout and drop stale build artifacts.

+++ src/drivers/uart.h
@@ -22,7 +22,7 @@ void uart_printf(char *string, ...);
 char uart_getchar();
 void read_line(char *buf, int size);
 int strcmp(const char *a, const char *b);
-int strncmp(const char *a, const char *b, int n);
+int strncmp(const char *a, const char *b, size_t n);
 size_t strlen(const char *s);
 char *strchr(const char *s, int c);
 int parse_two_args(char *input, char **arg1, char **arg2);
